Make stream iterators thread safe.
The access of stream iterators is not thread safe, and naively following
the pattern of taking the media src lock to access the streams will cause
deadlocks with gstreamer also occassionally taking this lock.

Factor out the stream iterators and make them all use the streamLock, which
seemed designed for this purpose but wasn't being consistently used.

Also remove the streamCondition, which isn't used anywhere.