Try to play AVC codec even if H.264 decoder only advertises byte-stream profile.
The playbin figures out that a conversion is needed automatically, so there's
no point special casing this here. Also remove the std::array, it's irrating to
update (two things to change, the length of the array and the thing to add) and
adds nothing over a stack array in this case.

This fixes the EncryptedEventData test in the YouTube 2018 EME conformance suite.

* platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.cpp:
(WebCore::codecSet):