[EME][NonWorking] Changed the events handling logic
The idea is to handle the events both in the demuxer thread or main
thread when the decryptor sends the event back to the pipeline if it
hadn't been handled before. This code attempts to fix that but threads
can mess things up.

In this case, when we get the message from the decryptor, we check if
it was reported to JS or not. If it wasn't we should send it (code
pending to add) and if it was already, we should dispatch the key. For
that we need to have the session mapped and that should happen during
attemptToDecryptWithLocalInstance but threads behave in other fashion
and the code to cache the session is run after the cache is accessed,
unleashing hell and dragons.