[EME][KindaWorking] Check for the validity of the session
Before attempting to decrypt we need to check for the session validity
because session can be set but be still invalid.

This makes eme-v3.html work but solution is not optimal because it is
working because the timeout of the decryptor is running out and we
need to find a way to wake it before. Maybe the solution is delaying
setting the session until it is valid.

MSE code is still broken.