[EME][MSE] Adapt to the right events management in Base
MediaPlayerPrivateGStreamerBase does all the events, sessions and init
datas management for us so we can get rid of most of the code. The
only important thing is properly dispatching the sessions to the
corresponding decryptors in the AppendPipelines.

The was an issue with the append states when shipping the sessions and
I realized that it was because of the KeyNegotiation state that is not
needed anymore as decryptors handle their sessions and wait for
decryption by themselves so I removed it.